Document bp circuit
