+++ crypto/src/bp_circuits/cloak.rs
@@ -1,72 +1,75 @@
-/*
-* This file implements the Cloak protocol for multiple confidential asset transactions.
-* The protocol proves that
-* 1) output amounts are in [0..2^64-1],
-* 2) for each asset type, input amount equals output amount
-*
-* Prover first sorts by inputs and outputs by asset types, and proves that sorted lists are a permutation
-* of input and output respectively. Then the sorted list are "merged", for each asset type we add up its amount into
-* a single record, zeroing-out the rest. Then, the merged input or merged output list is padded
-* with zero values so that they are of the same length.
-* Having same length, do a third permutation proof, proving that input and output are equal once
-* values of the same type are added.
-* Also, we append a range proof on each output value.
-*
-* Soundness note: Note that the prover may produce any permutation, and not necessarily a sorted one.
-* This doesn't help him as he
-* 1) permutation ensures he is not faking on any value.
-* 2) The "merge" proof ensures that only same asset types amounts are added
+//! This file implements the Cloak protocol for multiple confidential asset transactions.
+//! The protocol proves that
+//! 1) output amounts are in [0..2^64-1],
+//! 2) for each asset type, input amount equals output amount
+//!
+//! Prover first sorts by inputs and outputs by asset types, and proves that sorted lists
+//! are a permutation of input and output respectively. Then the sorted list are "merged",
+//! for each asset type we add up its amount into a single record, zeroing-out the rest.
+//! Then, the merged input or merged output list is padded with zero values so that they
+//! are of the same length.
+//! Having same length, do a third permutation proof, proving that input and output are
+//! equal once values of the same type are added.
+//! Also, we append a range proof on each output value.
+//!
+//! Soundness note: Note that the prover may produce any permutation, and not necessarily
+//! a sorted one.
+//! This doesn't help him as he
+//! 1) permutation ensures he is not faking on any value.
+//! 2) The "merge" proof ensures that only same asset types amounts are added
 
-* Permutation proof:
-* 1) For each record, (amount, asset type) produce a single variable (amount  + r * asset_type).
-* 2) To prove that a list is a permutation of another list, we build a circuit that test if the product of the values
-match on each list. On each product the elements are shifted by a random challenge.
-
-* Merge proof:
-* Given a sorted list of length n, the proof proceed by looking at consecutive pairs in1, in2
-* proving that respective output out1 and out2 are either
-*  + the same as in1 and in2, or
-*  +  - in1 and in2 have the same asset type, and
-      - out 1 is zero, and
-      - out2 asset type matched inputs asset type
-      - out2 amounts is the addition of the input amounts.
-* To have a final output list with all amounts aggregates by asset type,
-* the prover produces an intermediate list of length l - 2 storing the second output out2 of the previous description
-* Hence, each intermediate value is used as in1, and each input (except first) is used as in2.
-* Also, each intermediate value is used as out2, and each output is used as out1.
-* Notes:
-  - The fist in1 is actually input[0]
-  - The last intermediate value is appended to the output list.
-*/
+//! Permutation proof:
+//! 1) For each record, (amount, asset type) produce a single variable (amount  + r * asset_type).
+//! 2) To prove that a list is a permutation of another list, we build a circuit that test
+//! if the product of the values match on each list. On each product the elements are shifted
+//! by a random challenge.
+//!
+//! Merge proof:
+//! Given a sorted list of length n, the proof proceed by looking at consecutive pairs in1, in2
+//! proving that respective output out1 and out2 are either
+//!  - the same as in1 and in2, or
+//!  - in1 and in2 have the same asset type, and
+//!    - out 1 is zero, and
+//!    - out2 asset type matched inputs asset type
+//!    - out2 amounts is the addition of the input amounts.
+//! To have a final output list with all amounts aggregates by asset type,
+//! the prover produces an intermediate list of length l - 2 storing the second output out2
+//! of the previous description
+//! Hence, each intermediate value is used as in1, and each input (except first) is used as in2.
+//! Also, each intermediate value is used as out2, and each output is used as out1.
+//! Notes:
+//!  - The fist in1 is actually input[0]
+//!  - The last intermediate value is appended to the output list.
 
 use crate::basic::ristretto_pedersen_comm::RistrettoPedersenCommitment;
 use bulletproofs::r1cs::{
-    ConstraintSystem, Prover, RandomizableConstraintSystem, Variable, Verifier,
+    ConstraintSystem, LinearCombination, Prover, R1CSError, RandomizableConstraintSystem,
+    RandomizedConstraintSystem, Variable, Verifier,
 };
 use merlin::Transcript;
-use ruc::*;
-use zei_algebra::errors::ZeiError;
+use std::iter;
 use zei_algebra::{
-    ops::*,
+    prelude::*,
     ristretto::{CompressedRistretto, RistrettoScalar as Scalar},
-    Zero,
+    traits::Scalar as _,
 };
 
 /// Represent AssetRecord amount and asset type
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
 pub struct CloakValue {
+    /// Represent the amount
     pub amount: Scalar,
+    /// Represent the asset type
     pub asset_type: Scalar,
 }
 
 impl CloakValue {
+    /// Create an AssetRecord
     pub fn new(amount: Scalar, asset_type: Scalar) -> CloakValue {
         CloakValue { amount, asset_type }
     }
-}
 
-impl CloakValue {
-    /// Prover commits to amount and asset type using user provided blinding factors
+    /// Prover commits to amount and asset type using user provided blinding factors.
     /// Returns commitments and circuit variables
     pub fn commit_prover(
         &self,
@@ -87,6 +90,8 @@ impl CloakValue {
             },
         )
     }
+
+    /// Commit to amount and asset type using user provided blinding factors.
     pub fn commit(
         &self,
         pc_gens: &RistrettoPedersenCommitment,
@@ -104,14 +109,18 @@ impl CloakValue {
 /// Represent a CloakValue variable in the circuit
 #[derive(Clone, Copy, Debug)]
 pub struct CloakVariable {
+    /// Represent the amount variable
     pub amount: Variable,
+    /// Represent the asset type variable
     pub asset_type: Variable,
 }
 
 /// Represent a commitment Cloak value.
 #[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq)]
 pub struct CloakCommitment {
+    /// Represent the amount commitment
     pub amount: CompressedRistretto,
+    /// Represent the asset type commitment
     pub asset_type: CompressedRistretto,
 }
 
@@ -160,12 +169,12 @@ pub fn cloak<CS: RandomizableConstraintSystem>(
     }
 
     // do a proof of shuffle
-    n_gates += super::gadgets::cloak_shuffle_gadget(cs, merged_input_vars, merged_output_vars)
+    n_gates += cloak_shuffle_gadget(cs, merged_input_vars, merged_output_vars)
         .c(d!(ZeiError::R1CSProofError))?;
 
     // final range proof:
     for (i, out) in output_vars.iter().enumerate() {
-        n_gates += super::gadgets::range_proof_64(
+        n_gates += range_proof_64(
             cs,
             out.amount.into(),
             output_values.map(|out_values| out_values[i].amount),
@@ -176,7 +185,7 @@ pub fn cloak<CS: RandomizableConstraintSystem>(
     Ok(n_gates)
 }
 
-// Pad list with with value to expected len
+/// Pad list with with value to expected len
 fn pad<CS: ConstraintSystem>(
     cs: &mut CS,
     expected_len: usize,
@@ -192,7 +201,7 @@ fn pad<CS: ConstraintSystem>(
     Ok(())
 }
 
-// sorts the input list by asset_type
+/// sorts the input list by asset_type
 fn sort(input: &[CloakValue]) -> Vec<CloakValue> {
     let mut sorted_values = input.to_vec();
     let mut i = 0;
@@ -213,15 +222,15 @@ fn sort(input: &[CloakValue]) -> Vec<CloakValue> {
     sorted_values
 }
 
-// Aggregate amounts of consecutive CloakValues if asset_types match
-// Return vec 0 correspond to intermediate values storing the result
-// of each current pair. That is, intermediate[i] = Merge(intermediate[i-1], input[i])[1]
-// Where intermediate[0] = input[0]
-// Return vec 1 is the final aggregated values. That is output[i] = Merge(intermediate[i-1], input[i])[0]
-// Where output[l-1] = intermediate[l-1];
-// Intermediate list is of length (l-2) ( all intermediate produced except last)
-// Output list is of length l (all outputs plus last intermediate)
-pub(super) fn merge(sorted: &[CloakValue]) -> (Vec<CloakValue>, Vec<CloakValue>) {
+/// Aggregate amounts of consecutive CloakValues if asset_types match
+/// Return vec 0 correspond to intermediate values storing the result
+/// of each current pair. That is, intermediate[i] = Merge(intermediate[i-1], input[i])[1]
+/// Where intermediate[0] = input[0]. Return vec 1 is the final aggregated values.
+/// That is output[i] = Merge(intermediate[i-1], input[i])[0]
+/// Where output[l-1] = intermediate[l-1];
+/// Intermediate list is of length (l-2) ( all intermediate produced except last)
+/// Output list is of length l (all outputs plus last intermediate)
+fn merge(sorted: &[CloakValue]) -> (Vec<CloakValue>, Vec<CloakValue>) {
     let mut intermediate = vec![];
     let mut merged = vec![];
     let mut prev = sorted[0];
@@ -250,8 +259,8 @@ pub(super) fn merge(sorted: &[CloakValue]) -> (Vec<CloakValue>, Vec<CloakValue>)
     (intermediate, merged)
 }
 
-// Implements the sort and merge steps of the Cloak protocol
-pub(super) fn sort_and_merge<CS: RandomizableConstraintSystem>(
+/// Implements the sort and merge steps of the Cloak protocol
+fn sort_and_merge<CS: RandomizableConstraintSystem>(
     cs: &mut CS,
     vars: &[CloakVariable],
     values: Option<&[CloakValue]>,
@@ -280,16 +289,16 @@ pub(super) fn sort_and_merge<CS: RandomizableConstraintSystem>(
     let merged_vars =
         allocate_cloak_vector(cs, merged_values.as_ref().map(|(_, merged)| merged), len).c(d!())?;
 
-    n_gates += super::gadgets::cloak_shuffle_gadget(cs, vars.to_vec(), sorted_vars.clone())
+    n_gates += cloak_shuffle_gadget(cs, vars.to_vec(), sorted_vars.clone())
         .c(d!(ZeiError::R1CSProofError))?;
-    n_gates +=
-        super::gadgets::cloak_merge_gadget(cs, &sorted_vars, &intermediate_vars, &merged_vars)
-            .c(d!(ZeiError::R1CSProofError))?;
+    n_gates += cloak_merge_gadget(cs, &sorted_vars, &intermediate_vars, &merged_vars)
+        .c(d!(ZeiError::R1CSProofError))?;
 
     Ok((n_gates, merged_vars))
 }
 
-pub(crate) fn allocate_cloak_vector<CS: ConstraintSystem>(
+/// Batch alloc cloak value
+fn allocate_cloak_vector<CS: ConstraintSystem>(
     cs: &mut CS,
     list: Option<&Vec<CloakValue>>,
     len: usize,
@@ -322,19 +331,391 @@ pub(crate) fn allocate_cloak_vector<CS: ConstraintSystem>(
     })
 }
 
+/// Implements the merge gadget of the Cloak protocol
+fn cloak_merge_gadget<CS: RandomizableConstraintSystem>(
+    cs: &mut CS,
+    sorted: &[CloakVariable],
+    intermediate: &[CloakVariable],
+    merged: &[CloakVariable],
+) -> Result<usize> {
+    let in1 = sorted[0];
+    let out1 = merged[0];
+    let mut n_gates = 0;
+    let l = sorted.len();
+    if l == 0 {
+        return Ok(0);
+    }
+    if l == 1 {
+        // sorted and merges should be the same
+        cs.constrain(in1.amount - out1.amount);
+        cs.constrain(in1.asset_type - out1.asset_type);
+        return Ok(0);
+    }
+    assert_eq!(l, merged.len());
+    assert_eq!(l, intermediate.len() + 2);
+    let first_in = sorted[0];
+    let in1iter = iter::once(&first_in).chain(intermediate.iter());
+    let in2iter = sorted[1..l].iter();
+    let out1iter = merged[0..l - 1].iter();
+    let out2iter = intermediate.iter().chain(iter::once(&merged[l - 1]));
+
+    for (((in1, in2), out1), out2) in in1iter.zip(in2iter).zip(out1iter).zip(out2iter) {
+        n_gates += gate_mix(cs, *in1, *in2, *out1, *out2).c(d!())?;
+    }
+
+    Ok(n_gates)
+}
+
+/// I implement the mix gate gadget, proving that either
+/// in1 = out1 and in2 = out2  if in1 and in2 have different types or
+/// out1 = 0 and out2 = in1 + in2 if in1 and in2 have same type
+/// I return the number of left wires created
+fn gate_mix<CS: RandomizableConstraintSystem>(
+    cs: &mut CS,
+    in1: CloakVariable,
+    in2: CloakVariable,
+    out1: CloakVariable,
+    out2: CloakVariable,
+) -> Result<usize> {
+    cs.specify_randomized_constraints(move |cs| {
+        let w1 = cs.challenge_scalar(b"mix challenge1");
+        let w2 = cs.challenge_scalar(b"mix challenge2");
+        let w3 = cs.challenge_scalar(b"mix challenge3");
+        let (_, _, out) = cs.multiply(
+            (in1.amount - out1.amount) +          // quantity maintains
+                (in1.asset_type - out1.asset_type) * w1 + // asset type maintains in first input
+                (in2.amount - out2.amount) * w2 + // quantity maintains
+                (in2.asset_type - out2.asset_type) * w3, // asset type maintains in second input
+            out1.amount + // or out1 is 0
+                (in1.asset_type - in2.asset_type) * w1 + // or in flavors are equal
+                (out2.amount - in1.amount - in2.amount) * w2 // or out2 is the sum of the inputs
+                + (out2.asset_type - in1.asset_type) * w3, // in 1 and out2 have same asset type
+        );
+        cs.constrain(out.into());
+        Ok(())
+    })
+    .c(d!())?;
+    Ok(1usize)
+}
+
+/// Prove shuffling of a list of CloakValues
+/// Return the number of gates created
+fn cloak_shuffle_gadget<CS: RandomizableConstraintSystem>(
+    cs: &mut CS,
+    input: Vec<CloakVariable>,
+    permuted: Vec<CloakVariable>,
+) -> Result<usize> {
+    let l = input.len();
+    if l != permuted.len() {
+        return Err(eg!(R1CSError::GadgetError {
+            description: "list shuffle error, input and output list differ in length".to_string(),
+        }));
+    }
+    if l == 0 {
+        return Ok(0usize);
+    }
+    if l == 1 {
+        cs.constrain(permuted[0].amount - input[0].amount);
+        cs.constrain(permuted[0].asset_type - input[0].asset_type);
+        return Ok(0usize);
+    }
+
+    cs.specify_randomized_constraints(move |cs| {
+        let challenge = cs.challenge_scalar(b"k-value shuffle challenge");
+        let mut single_input = Vec::with_capacity(l);
+        let mut single_perm = Vec::with_capacity(l);
+
+        for (in_var, perm_var) in input.iter().zip(permuted.iter()) {
+            //compute a single representative for the pair
+            let (single_in, single_pe, _) = cs.multiply(
+                in_var.amount + challenge * in_var.asset_type,
+                perm_var.amount + challenge * perm_var.asset_type,
+            );
+            single_input.push(single_in);
+            single_perm.push(single_pe);
+        }
+
+        list_shuffle(cs, &single_input[..], &single_perm[..])
+            .c(d!())
+            .map_err(|e| R1CSError::GadgetError {
+                description: e.to_string(),
+            })
+            .map(|_| ())
+    })
+    .c(d!())?;
+
+    // list_shuffle does 2*(l-1) multiplications
+    Ok(l + 2 * (l - 1))
+}
+
+/// Prove shuffling of a list of values
+/// Return the number of left wires created
+fn list_shuffle<CS: RandomizedConstraintSystem>(
+    cs: &mut CS,
+    input: &[Variable],
+    permuted: &[Variable],
+) -> Result<usize> {
+    let l = input.len();
+    if l != permuted.len() {
+        return Err(eg!(R1CSError::GadgetError {
+            description: "list shuffle error, input and output list differ in length".to_string(),
+        }));
+    }
+    if l == 0 {
+        return Ok(0usize);
+    }
+    if l == 1 {
+        cs.constrain(permuted[0] - input[0]);
+        return Ok(0usize);
+    }
+
+    let challenge = cs.challenge_scalar(b"shuffle challenge");
+
+    // Make last x multiplier for i = l-1 and l-2
+    let (_, _, last_mulx_out) = cs.multiply(input[l - 1] - challenge, input[l - 2] - challenge);
+
+    // Make multipliers for x from i == [0, l-3]
+    let first_mulx_out = (0..l - 2).rev().fold(last_mulx_out, |prev_out, i| {
+        let (_, _, o) = cs.multiply(prev_out.into(), input[i] - challenge);
+        o
+    });
+
+    // Make last y multiplier for i = l-1 and l-2
+    let (_, _, last_muly_out) =
+        cs.multiply(permuted[l - 1] - challenge, permuted[l - 2] - challenge);
+
+    // Make multipliers for y from i == [0, l-3]
+    let first_muly_out = (0..l - 2).rev().fold(last_muly_out, |prev_out, i| {
+        let (_, _, o) = cs.multiply(prev_out.into(), permuted[i] - challenge);
+        o
+    });
+
+    // Constrain last x mul output and last y mul output to be equal
+    cs.constrain(first_mulx_out - first_muly_out);
+
+    // l-1 multiplications for input + l-1 multiplication for permuted
+    Ok(2 * (l - 1))
+}
+
+/// I prove that value is in [0..2^64-1]
+fn range_proof_64<CS: ConstraintSystem>(
+    cs: &mut CS,
+    mut var: LinearCombination,
+    value: Option<Scalar>,
+) -> Result<usize> {
+    let mut exp_2 = Scalar::one();
+    let n_usize = 64usize;
+    let value_bytes = value.as_ref().map(|v| v.to_bytes());
+    for i in 0..n_usize {
+        // Create low-level variables and add them to constraints
+        let (a, b, o) = match value_bytes.as_ref() {
+            Some(bytes) => {
+                let index = i >> 3;
+                if index > bytes.len() {
+                    // This could happen due to the scalar's representation
+                    return Err(eg!(R1CSError::FormatError));
+                }
+                let bit = ((bytes[index] >> (i & 7)) & 1u8) as i8;
+                let assignment = (Scalar::from(1 - bit as u32), Scalar::from(bit as u32));
+                cs.allocate_multiplier(Some(assignment).map(|(a, b)| (a.0, b.0)))
+            }
+            None => cs.allocate_multiplier(None),
+        }
+        .c(d!())?;
+
+        // Enforce a * b = 0, so one of (a,b) is zero
+        cs.constrain(o.into());
+
+        // Enforce that a = 1 - b, so they both are 1 or 0.
+        cs.constrain(a + (b - 1u64));
+
+        // Add `-b_i*2^i` to the linear combination
+        // in order to form the following constraint by the end of the loop:
+        // v = Sum(b_i * 2^i, i = 0..n-1)
+        var = var - b * exp_2.0;
+        exp_2 = exp_2.add(&exp_2);
+    }
+    // Enforce that v = Sum(b_i * 2^i, i = 0..n-1)
+    cs.constrain(var);
+
+    // one multiplication gate per bit
+    Ok(n_usize)
+}
+
 #[cfg(test)]
 pub mod tests {
-    use crate::bp_circuits::cloak::{CloakCommitment, CloakValue};
-    use bulletproofs::r1cs::{Prover, R1CSProof, Verifier};
-    use bulletproofs::{BulletproofGens, PedersenGens};
-    use itertools::Itertools;
+    use crate::bp_circuits::cloak::{
+        allocate_cloak_vector, CloakCommitment, CloakValue, CloakVariable,
+    };
+    use bulletproofs::{
+        r1cs::{Prover, R1CSProof, Verifier},
+        BulletproofGens, PedersenGens,
+    };
     use lazy_static::lazy_static;
     use merlin::Transcript;
     use rand_chacha::ChaChaRng;
-    use rand_core::SeedableRng;
-    use zei_algebra::{ops::*, ristretto::RistrettoScalar, traits::Scalar, Zero};
+    use zei_algebra::{prelude::*, ristretto::RistrettoScalar};
 
-    // Taken from https://github.com/stellar/slingshot/tree/main/cloak
+    #[test]
+    fn test_cloak_merge() {
+        let pc_gens = PedersenGens::default();
+
+        let sorted_values = vec![
+            CloakValue::new(RistrettoScalar::from(1u32), RistrettoScalar::from(10u32)),
+            CloakValue::new(RistrettoScalar::from(3u32), RistrettoScalar::from(10u32)),
+            CloakValue::new(RistrettoScalar::from(2u32), RistrettoScalar::from(11u32)),
+            CloakValue::new(RistrettoScalar::from(5u32), RistrettoScalar::from(11u32)),
+            CloakValue::new(RistrettoScalar::from(4u32), RistrettoScalar::from(12u32)),
+        ];
+
+        let mid_values = vec![
+            CloakValue::new(RistrettoScalar::from(4u32), RistrettoScalar::from(10u32)),
+            CloakValue::new(RistrettoScalar::from(2u32), RistrettoScalar::from(11u32)),
+            CloakValue::new(RistrettoScalar::from(7u32), RistrettoScalar::from(11u32)),
+        ];
+        let out_values = vec![
+            CloakValue::new(RistrettoScalar::zero(), RistrettoScalar::from(10u32)),
+            CloakValue::new(RistrettoScalar::from(4u32), RistrettoScalar::from(10u32)),
+            CloakValue::new(RistrettoScalar::zero(), RistrettoScalar::from(11u32)),
+            CloakValue::new(RistrettoScalar::from(7u32), RistrettoScalar::from(11u32)),
+            CloakValue::new(RistrettoScalar::from(4u32), RistrettoScalar::from(12u32)),
+        ];
+
+        let mut prover_transcript = Transcript::new(b"test");
+        let mut prover = Prover::new(&pc_gens, &mut prover_transcript);
+        let sorted =
+            allocate_cloak_vector(&mut prover, Some(&sorted_values), sorted_values.len()).unwrap();
+        let mid = allocate_cloak_vector(&mut prover, Some(&mid_values), mid_values.len()).unwrap();
+        let added =
+            allocate_cloak_vector(&mut prover, Some(&out_values), out_values.len()).unwrap();
+        let num_wires =
+            super::cloak_merge_gadget(&mut prover, &sorted[..], &mid[..], &added[..]).unwrap();
+        let bp_gens = BulletproofGens::new(
+            (num_wires + 2 * (sorted.len() + mid.len() + added.len())).next_power_of_two(),
+            1,
+        );
+        let proof = prover.prove(&bp_gens).unwrap();
+
+        let mut verifier_transcript = Transcript::new(b"test");
+        let mut verifier = Verifier::new(&mut verifier_transcript);
+        let sorted = allocate_cloak_vector(&mut verifier, None, sorted_values.len()).unwrap();
+        let mid = allocate_cloak_vector(&mut verifier, None, mid_values.len()).unwrap();
+        let added = allocate_cloak_vector(&mut verifier, None, out_values.len()).unwrap();
+        let num_wires =
+            super::cloak_merge_gadget(&mut verifier, &sorted[..], &mid[..], &added[..]).unwrap();
+        let bp_gens = BulletproofGens::new(
+            (num_wires + 2 * (sorted.len() + mid.len() + added.len())).next_power_of_two(),
+            1,
+        );
+        assert!(verifier.verify(&proof, &pc_gens, &bp_gens).is_ok());
+
+        // test the same using commitments
+        let mut prover_transcript = Transcript::new(b"test");
+        let mut prover = Prover::new(&pc_gens, &mut prover_transcript);
+        let sorted_coms_vars: Vec<(CloakCommitment, CloakVariable)> = sorted_values
+            .iter()
+            .map(|value| {
+                value.commit_prover(
+                    &mut prover,
+                    &CloakValue::new(RistrettoScalar::from(1u32), RistrettoScalar::from(2u32)),
+                )
+            })
+            .collect();
+        let mid = allocate_cloak_vector(&mut prover, Some(&mid_values), mid_values.len()).unwrap();
+        let added =
+            allocate_cloak_vector(&mut prover, Some(&out_values), out_values.len()).unwrap();
+        let sorted_vars: Vec<CloakVariable> =
+            sorted_coms_vars.iter().map(|(_, var)| *var).collect();
+        let num_wires =
+            super::cloak_merge_gadget(&mut prover, &sorted_vars[..], &mid[..], &added[..]).unwrap();
+        let num_wires = num_wires + added.len() + mid.len();
+        let bp_gens = BulletproofGens::new(num_wires.next_power_of_two(), 1);
+        let proof = prover.prove(&bp_gens).unwrap();
+
+        let mut verifier_transcript = Transcript::new(b"test");
+        let mut verifier = Verifier::new(&mut verifier_transcript);
+        let sorted_coms: Vec<CloakCommitment> =
+            sorted_coms_vars.iter().map(|(com, _)| (*com)).collect();
+        let sorted_vars: Vec<CloakVariable> = sorted_coms
+            .iter()
+            .map(|com| com.commit_verifier(&mut verifier))
+            .collect();
+        let mid = allocate_cloak_vector(&mut verifier, None, mid_values.len()).unwrap();
+        let added = allocate_cloak_vector(&mut verifier, None, out_values.len()).unwrap();
+        let num_wires =
+            super::cloak_merge_gadget(&mut verifier, &sorted_vars[..], &mid[..], &added[..])
+                .unwrap();
+        let num_wires = num_wires + added.len() + mid.len();
+        let bp_gens = BulletproofGens::new(num_wires.next_power_of_two(), 1);
+        assert!(verifier.verify(&proof, &pc_gens, &bp_gens).is_ok());
+    }
+
+    #[test]
+    fn test_shuffle() {
+        let pc_gens = PedersenGens::default();
+        let input_values = vec![
+            CloakValue::new(RistrettoScalar::from(10u32), RistrettoScalar::from(10u32)),
+            CloakValue::new(RistrettoScalar::from(20u32), RistrettoScalar::from(20u32)),
+            CloakValue::new(RistrettoScalar::from(30u32), RistrettoScalar::from(30u32)),
+            CloakValue::new(RistrettoScalar::from(40u32), RistrettoScalar::from(40u32)),
+            CloakValue::new(RistrettoScalar::from(50u32), RistrettoScalar::from(50u32)),
+        ];
+
+        let shuffled_values = vec![
+            CloakValue::new(RistrettoScalar::from(20u32), RistrettoScalar::from(20u32)),
+            CloakValue::new(RistrettoScalar::from(40u32), RistrettoScalar::from(40u32)),
+            CloakValue::new(RistrettoScalar::from(10u32), RistrettoScalar::from(10u32)),
+            CloakValue::new(RistrettoScalar::from(50u32), RistrettoScalar::from(50u32)),
+            CloakValue::new(RistrettoScalar::from(30u32), RistrettoScalar::from(30u32)),
+        ];
+
+        let mut prover_transcript = Transcript::new(b"test");
+        let mut prover = Prover::new(&pc_gens, &mut prover_transcript);
+        let input =
+            allocate_cloak_vector(&mut prover, Some(&input_values), input_values.len()).unwrap();
+        let shuffled =
+            allocate_cloak_vector(&mut prover, Some(&shuffled_values), shuffled_values.len())
+                .unwrap();
+        let num_wires =
+            super::cloak_shuffle_gadget(&mut prover, input.to_vec(), shuffled.to_vec()).unwrap();
+        let num_wires = num_wires + input.len() + shuffled.len();
+        let bp_gens = BulletproofGens::new(num_wires.next_power_of_two(), 1);
+        let proof = prover.prove(&bp_gens).unwrap();
+
+        let mut verifier_transcript = Transcript::new(b"test");
+        let mut verifier = Verifier::new(&mut verifier_transcript);
+        let input = allocate_cloak_vector(&mut verifier, None, input_values.len()).unwrap();
+        let shuffled = allocate_cloak_vector(&mut verifier, None, shuffled_values.len()).unwrap();
+        super::cloak_shuffle_gadget(&mut verifier, input.to_vec(), shuffled.to_vec()).unwrap();
+        assert!(verifier.verify(&proof, &pc_gens, &bp_gens).is_ok());
+
+        let bad_shuffle_values = vec![
+            CloakValue::new(RistrettoScalar::zero(), RistrettoScalar::zero()),
+            CloakValue::new(RistrettoScalar::from(40u32), RistrettoScalar::from(40u32)),
+            CloakValue::new(RistrettoScalar::from(10u32), RistrettoScalar::from(10u32)),
+            CloakValue::new(RistrettoScalar::from(50u32), RistrettoScalar::from(50u32)),
+            CloakValue::new(RistrettoScalar::from(30u32), RistrettoScalar::from(30u32)),
+        ];
+
+        let mut prover_transcript = Transcript::new(b"test");
+        let mut prover = Prover::new(&pc_gens, &mut prover_transcript);
+        let input = allocate_cloak_vector(&mut prover, Some(&input_values), 0).unwrap();
+        let bad_shuffle = allocate_cloak_vector(&mut prover, Some(&bad_shuffle_values), 0).unwrap();
+        let num_wires =
+            super::cloak_shuffle_gadget(&mut prover, input.to_vec(), bad_shuffle).unwrap();
+        let num_wires = num_wires + input.len() + shuffled.len();
+        let bp_gens = BulletproofGens::new(num_wires.next_power_of_two(), 1);
+        let proof = prover.prove(&bp_gens).unwrap();
+
+        let mut verifier_transcript = Transcript::new(b"test");
+        let mut verifier = Verifier::new(&mut verifier_transcript);
+        let input = allocate_cloak_vector(&mut verifier, None, input_values.len()).unwrap();
+        let bad_shuffle =
+            allocate_cloak_vector(&mut verifier, None, bad_shuffle_values.len()).unwrap();
+        super::cloak_shuffle_gadget(&mut verifier, input.to_vec(), bad_shuffle.to_vec()).unwrap();
+        assert!(verifier.verify(&proof, &pc_gens, &bp_gens).is_err());
+    }
+
     fn yuan(q: u64) -> CloakValue {
         CloakValue {
             amount: RistrettoScalar::from(q),
@@ -518,9 +899,9 @@ pub mod tests {
         );
     }
 
-    ////////////////////////////////////////////////////////////////////////////////////////////////
-    // Port of https://github.com/stellar/slingshot/blob/main/cloak/tests/cloak.rs        //
-    ////////////////////////////////////////////////////////////////////////////////////////////////
+    /////////////////////////////////////////////////////////////////////////////////////////
+    // Port of https://github.com/stellar/slingshot/blob/main/spacesuit/tests/spacesuit.rs //
+    /////////////////////////////////////////////////////////////////////////////////////////
 
     // m=1, n=1
     #[test]
