+++ crypto/src/bp_circuits/mod.rs
@@ -1,3 +1,2 @@
 pub mod cloak;
-mod gadgets;
 pub mod range_proofs;
